various changes to clean warnings from MSVC builds
git-svn-id: https://svn.osgeo.org/gdal/trunk@21680 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/gsg/gs7bgdataset.cpp
@@ -503,8 +503,7 @@ GDALDataset *GS7BGDataset::Open( GDALOpenInfo * poOpenInfo )
         return NULL;
     }
 
-    poDS->nData_Position =  VSIFTellL(poDS->fp);
-
+    poDS->nData_Position =  (size_t) VSIFTellL(poDS->fp);
 
     /* --------------------------------------------------------------------*/
     /*      Initialize any PAM information.                                */
